Replace int to double

+++ src/Week_02_Point_Pen/Point3D.java
@@ -4,21 +4,21 @@ package Week_02_Point_Pen;
 public class Point3D extends Point2D {
 
     // Additional coordinate for the third dimension (z-axis)
-    private int z;
+    private double z;
 
     // Constructor to initialize x, y (from Point2D) and z
-    public Point3D(int x, int y, int z) {
+    public Point3D(double x, double y, double z) {
         super(x, y);  // Call parent constructor to set x and y
         this.z = z;
     }
 
     // Getter method to access z-coordinate
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 
     // Setter method to update z-coordinate
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
